seperated server and webserver
